+++ bean-searcher/src/main/java/cn/zhxu/bs/filter/ArrayValueParamFilter.java
@@ -17,7 +17,15 @@ import java.util.Objects;
  */
 public class ArrayValueParamFilter implements ParamFilter {
 
-    private String separator = "-";
+    private final String separator;
+
+    public ArrayValueParamFilter() {
+        this("-");
+    }
+
+    public ArrayValueParamFilter(String separator) {
+        this.separator = Objects.requireNonNull(separator);
+    }
 
     @Override
     public <T> Map<String, Object> doFilter(BeanMeta<T> beanMeta, Map<String, Object> paraMap)
@@ -43,10 +51,6 @@ public class ArrayValueParamFilter implements ParamFilter {
 
     public String getSeparator() {
         return separator;
-    }
-
-    public void setSeparator(String separator) {
-        this.separator = Objects.requireNonNull(separator);
     }
 
 }